Renamed the runner modules.
