+++ crates/bld_runner/src/runner/mod.rs
@@ -1,2 +1,2 @@
-pub mod version1;
-pub mod version2;
+pub mod v1;
+pub mod v2;
