+++ crates/bld_runner/src/sync/versioned.rs
@@ -1,17 +1,17 @@
-use crate::runner::version1;
-use crate::runner::version2;
+use crate::runner::v1;
+use crate::runner::v2;
 use anyhow::Result;
 
 pub enum VersionedRunner {
-    Version1(version1::Runner),
-    Version2(version2::Runner),
+    V1(v1::Runner),
+    V2(v2::Runner),
 }
 
 impl VersionedRunner {
     pub async fn run(self) -> Result<()> {
         match self {
-            Self::Version1(runner) => runner.run().await.await,
-            Self::Version2(runner) => runner.run().await.await,
+            Self::V1(runner) => runner.run().await.await,
+            Self::V2(runner) => runner.run().await.await,
         }
     }
 }
